refactor : remove services and add use cases
